Move pre retrieval predictors to own package.
