+++ src/main/java/lucene4ir/predictor/pre/AvgICTFQPPredictor.java
@@ -1,5 +1,6 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
+import lucene4ir.predictor.PreQPPredictor;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
 
@@ -7,35 +8,41 @@ import java.io.IOException;
 
 /**
  * Created by Harry Scells on 28/8/17.
+ * Averaged Inverse Collection Term Frequency (AvICTF)
+ * ICF = log(term_count / tf(q) )
  */
-public class QSQPPredictor extends QPPredictor {
+public class AvgICTFQPPredictor extends PreQPPredictor {
 
-    public QSQPPredictor(IndexReader ir) {
+    private double termCount;
+
+    public AvgICTFQPPredictor(IndexReader ir) {
         super(ir);
+        try {
+            termCount = getTermCount();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
     public double scoreQuery(String qno, Query q) {
-        // Number of documents containing at least one of the query terms.
-        double Nq = Double.POSITIVE_INFINITY;
         String[] termTuples = q.toString().split(" ");
+        double sumICF = 0.0;
+        double m = termTuples.length;
 
         for (String termTuple : termTuples) {
             String[] terms = termTuple.split(":");
             if (terms.length == 2) {
                 String term = terms[1];
                 try {
-                    double df = getDF(term);
-                    // What is the smallest possible df?
-                    if (df < Nq && df > 0) {
-                        Nq = df;
-                    }
+                    double tf = getTF(term);
+                    sumICF += log2(termCount) - log2(1 + tf);
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
         }
 
-        return -Math.log((1 + Nq) / (1 + reader.numDocs()));
+        return (1 / m) * sumICF;
     }
 }
