+++ src/main/java/lucene4ir/predictor/pre/CharLenQPPredictor.java
@@ -1,15 +1,16 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
+import lucene4ir.predictor.PreQPPredictor;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
 
 /**
  * Created by leif on 13/08/2017.
  */
-public class CharLenQPPredictor extends QPPredictor {
+public class CharLenQPPredictor extends PreQPPredictor {
 
 
-    public CharLenQPPredictor(IndexReader ir){
+    public CharLenQPPredictor(IndexReader ir) {
         super(ir);
     }
 
@@ -21,7 +22,7 @@ public class CharLenQPPredictor extends QPPredictor {
         int w = terms.length;
         int c = qstr.length();
 
-        return (double)(c - (w-1));
+        return (double) (c - (w - 1));
     }
 
 }
