+++ src/main/java/lucene4ir/predictor/pre/MaxSCQQPPredictor.java
@@ -1,4 +1,4 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
@@ -7,44 +7,35 @@ import java.io.IOException;
 
 /**
  * Created by Harry Scells on 28/8/17.
- * Summed Collection Query Similarity (SumSCQ)
- * sum over all t in q:
- * (1 +ln( cf(t))) * (1+N/df(t) )
+ * Maximum Collection Query Similarity MaxSCQ
  */
-public class SumSCQQPPredictor extends QPPredictor {
+public class MaxSCQQPPredictor extends SumSCQQPPredictor {
 
-    private double docCount;
-
-    public SumSCQQPPredictor(IndexReader ir) {
+    public MaxSCQQPPredictor(IndexReader ir) {
         super(ir);
-        docCount = reader.numDocs() + 1;
-    }
-
-    double calculateSCQ(String term) throws IOException {
-        double tf = getTF(term);
-        double df = getDF(term);
-        double idf = (docCount / df);
-        return (1 + Math.log(1 + tf)) * Math.log(1 + idf);
     }
 
     @Override
     public double scoreQuery(String qno, Query q) {
         String[] termTuples = q.toString().split(" ");
-        double sumSCQ = 0.0;
+        double maxSCQ = 0.0;
 
         for (String termTuple : termTuples) {
             String[] terms = termTuple.split(":");
             if (terms.length == 2) {
                 String term = terms[1];
                 try {
-                    sumSCQ += calculateSCQ(term);
+                    double SCQ = calculateSCQ(term);
+                    if (SCQ > maxSCQ) {
+                        maxSCQ = SCQ;
+                    }
+
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
         }
 
-        return sumSCQ;
+        return maxSCQ;
     }
-
 }
