+++ src/main/java/lucene4ir/predictor/pre/QSQPPredictor.java
@@ -1,5 +1,6 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
+import lucene4ir.predictor.PreQPPredictor;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
 
@@ -7,35 +8,35 @@ import java.io.IOException;
 
 /**
  * Created by Harry Scells on 28/8/17.
- * Maximum Collection Query Similarity MaxSCQ
  */
-public class MaxSCQQPPredictor extends SumSCQQPPredictor {
+public class QSQPPredictor extends PreQPPredictor {
 
-    public MaxSCQQPPredictor(IndexReader ir) {
+    public QSQPPredictor(IndexReader ir) {
         super(ir);
     }
 
     @Override
     public double scoreQuery(String qno, Query q) {
+        // Number of documents containing at least one of the query terms.
+        double Nq = Double.POSITIVE_INFINITY;
         String[] termTuples = q.toString().split(" ");
-        double maxSCQ = 0.0;
 
         for (String termTuple : termTuples) {
             String[] terms = termTuple.split(":");
             if (terms.length == 2) {
                 String term = terms[1];
                 try {
-                    double SCQ = calculateSCQ(term);
-                    if (SCQ > maxSCQ) {
-                        maxSCQ = SCQ;
+                    double df = getDF(term);
+                    // What is the smallest possible df?
+                    if (df < Nq && df > 0) {
+                        Nq = df;
                     }
-
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
         }
 
-        return maxSCQ;
+        return -Math.log((1 + Nq) / (1 + reader.numDocs()));
     }
 }
