+++ src/main/java/lucene4ir/predictor/pre/SCSQPPredictor.java
@@ -1,4 +1,4 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
