+++ src/main/java/lucene4ir/predictor/pre/StdDevIDFQPPredictor.java
@@ -1,5 +1,6 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
+import lucene4ir.predictor.PreQPPredictor;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.search.Query;
 
@@ -10,7 +11,7 @@ import java.util.List;
  * Created by Harry Scells on 28/8/17.
  * Standard Deviation of IDF (DevIDF)
  */
-public class StdDevIDFQPPredictor extends QPPredictor {
+public class StdDevIDFQPPredictor extends PreQPPredictor {
 
     public StdDevIDFQPPredictor(IndexReader ir) {
         super(ir);
