+++ src/main/java/lucene4ir/predictor/pre/TermLenQPPredictor.java
@@ -1,15 +1,15 @@
-package lucene4ir.predictor;
+package lucene4ir.predictor.pre;
 
+import lucene4ir.predictor.PreQPPredictor;
 import org.apache.lucene.index.IndexReader;
-import org.apache.lucene.queryparser.classic.QueryParser;
 import org.apache.lucene.search.Query;
 
 /**
  * Created by leif on 13/08/2017.
  */
-public class TermLenQPPredictor extends QPPredictor {
+public class TermLenQPPredictor extends PreQPPredictor {
 
-    public TermLenQPPredictor(IndexReader ir){
+    public TermLenQPPredictor(IndexReader ir) {
         super(ir);
     }
